Changed color of breadCrumbs on albumDetail
