+++ src/App.scss
@@ -5,6 +5,7 @@
 @use "./src/styles/photoDetail.scss";
 @use "./src/styles/arrowsStyles.scss";
 @use "./src/styles/navigationDrawer.scss";
+@use "./src/styles/breadCrumbs.scss";
 *,
 *::before,
 *:after {
