+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as AlbumDetail } from "./AlbumDetail";
 export { default as AlbumList } from "./AlbumList";
 export { default as AlbumThumb } from "./AlbumThumb";
-export { default as Breadcrumbs } from "./Breadcrumbs";
+export { default as BreadCrumbs } from "./BreadCrumbs";
 export { default as Footer } from "./Footer";
 export { default as Header } from "./Header";
 export { default as Photo } from "./Photo";
